Create ReweightDeformersTask
